Send debug push notifications directly from PushListenerService
Previously debug logs were sent via broadcast to ProcessingService, which auto-stops after 5s of idle. Logs were lost when the app wasn't open. Now debug logs are sent directly from PushListenerService (which is always alive as a NotificationListenerService) to the backend, ensuring they reach the server regardless of ProcessingService state.
